Functioning PB Code Tested Using Serial
Using the serial monitor, testing has shown that we can enable and
disable every bus, overcurrent is properly detected on every bus and bus
will turn off even if you command it to turn on again and again, and
current can be measured on each bus.  However, Pack Voltage meausing and
fan cntrl was not tested.