Voltage Measuring and Fan Cntrl Functioning
VOLTS_MAX was changed from 33.6 to 40 and now I recieve acurate pack
voltage measurments. I also changed the pin assignment for FAN_CNTRL
from 42 to 2 so now the fan can be turned on and off. Overall, power
board functions as desgined aside form the fact that the com and logic
busses cannot be individually turned off/on. It is only possible to turn
the entire com/logic regulator on/off. This is due to an error in the
board circuit and will be fixed for Rev2.