Communication with BMS to Base Station Tested
The communication of data from Bms to PB to base station and also
commands to bms from base station have been confirmed to be working.
This is working between just the bms's microcontroller, pb's Tiva, and
base station. This was not tested with the microcontrollers on the
actual boards. That is the next test to be performed. Otherwise, the
code was cleaned up and data ids were added.